Move is_commentable_cohorted to django_comment_client
